BE_220912_Community : 서비스impl 진행 중, controller - login API Rebase 후 완성 예정

+++ backend/src/main/java/com/ssafy/kirin/dto/request/CommunityCommentWriteDTO.java
@@ -1,4 +1,4 @@
 package com.ssafy.kirin.dto.request;
 
-public record CommunityCommentWriteDTO(String content, boolean isComment, int parentId) {
+public record CommunityCommentWriteDTO(String content, boolean isComment, long parentId) {
 }
